refactor: trocear EvaluationWizard en components/evaluation (Fase 1.9)
976 → 360 líneas. El wizard queda como orquestación; extraídos a
components/evaluation/: WelcomePhase, QuestionInput (input por tipo de
pregunta), SuccessPhase, WizardButton, ScaleInput y wizardStyles.ts (armazón).

Cero cambios de comportamiento (mismo JSX/estilos, callbacks por props). Se
mantiene el export nombrado EvaluationAnswers en el mismo path (lo importan
EvaluacionScreen/EvaluacionAppScreen/SurveyScreen). Gigantes >800: 12 → 11.
typecheck/typecheck:tests/lint(0 err)/test(197) ok.
